fixed merge conflict issue in database, began development of second page signup form
