+++ client/src/components/Profile.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Profile = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Profile
